Refactor: Move assign-roles endpoint to appropriate controller
