+++ src/main/java/com/activecourses/upwork/service/role/RoleServiceImpl.java
@@ -1,18 +1,26 @@
-package com.activecourses.upwork.service.user;
+package com.activecourses.upwork.service.role;
 
 import com.activecourses.upwork.model.Role;
+import com.activecourses.upwork.model.User;
 import com.activecourses.upwork.repository.role.RoleRepository;
+import com.activecourses.upwork.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class RoleServiceImpl implements RoleService {
 
     private final RoleRepository roleRepository;
+    private final UserRepository userRepository;
+    private static final Logger logger = LoggerFactory.getLogger(RoleServiceImpl.class);
 
     @Override
     public Role addRole(Role role) {
@@ -43,5 +51,19 @@ public class RoleServiceImpl implements RoleService {
     @Override
     public List<Role> getAllRoles() {
         return roleRepository.findAll();
+    }
+
+    @Override
+    public boolean assignRolesToUser(int userId, Map<String, Object> roles) {
+        logger.info("Assigning roles to user with id: {}", userId);
+        Optional<User> user = userRepository.findById(userId);
+        User unwrappedUser = user.orElseThrow(() -> new RuntimeException("User not found: " + userId));
+        List<Role> roleList = roles.keySet().stream()
+                .map(roleName -> roleRepository.findByName(roleName)
+                        .orElseThrow(() -> new RuntimeException("Role not found: " + roleName)))
+                .collect(Collectors.toList());
+        unwrappedUser.setRoles(roleList);
+        userRepository.save(unwrappedUser);
+        return true;
     }
 }
